Remove unused code from nodes component.
